requirements.txt
asgiref==3.8.1
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
diff-match-patch==20230430
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.2
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-import-export==4.1.1
python-dotenv==1.0.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
tablib==3.5.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
